requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
coverage==7.6.1
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.110.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pytest==8.2.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-randomly==3.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.29.0
websockets==13.0.1
